Added Supabase + Login Page
Added "Magic Link" based Login.

There are currently two URL groups. Guest & Logged in.
If not logged in and trying to access any page, the User will be redirected to the Login Page.
If logged in and trying to access a Guest page (Login), the User will be redirected to the homepage

The Magic Link method combines both Registration & Login. Emails are handled by Supabase itself.

---

Added a .env.example file.
All variables which are required to run the website are included in here. Values aren't included for security reasons.