Fixed User Management Alerts
Fixed the "Logged In" Alert only being raised for the SIGNED_IN Event.
When the Magic Link is clicked, a new Tab will be opened that raises the INITIAL_SESSION Event instead, even though the logged in session is now available.

If a Session is available and it isn't the already existing one, the "Logged In" Alert will be raised regardless of the Event type.

Due to the SIGNED_OUT Event also setting the session to null, the "Logged out" Alert has been moved.

---

supabase:auth will now only be invalidated if the Session changed. This happens after it was set via setSession.