Admin Feedback Page Base Layout
Added the Base Layout for the Admin Feedback Page.

The Promotion for which feedback should be displayed can be selected and the respective feedback will be loaded.
The Feedback Text & creation date is already being displayed, although this is missing any styling.
Promotions are loaded on the server, Feedback is loaded on the Client.
As only admins can visit these pages, there's not as much emphasis on everything happening on the Server as there was with the user-facing pages.

Some basic Pagination is available (10 entries per page).

---

Some minor cleanup in other files.
Removing unused variables and things like that.