Promotion Creation / Update Form
Added a Form for creating/updating Promotions.

When editing an existing Promotion, the fields are pre-filled with the values from the Promotion.
Keys can be added/removed. Changes are only applied when the "Save" Button is clicked.

Also fixed Promotions with no Expiration Date displaying "null". Now says "None".

---

Modal now allows the entire lower section (Buttons) to be overriden.